+++ packages/yumemi_lints/lib/src/models/project_type.dart
@@ -1,0 +1,4 @@
+enum ProjectType {
+  dart,
+  flutter,
+}